Make problem_api.c easier to comprehend
In my opinion this verison of the code is simpler.

Instead of having two kinds of callbacks, and *then* hardcoding the action
when they are successful (adding a directory name to the result list),
I combined all these actions into one callback.

This may allow simplification of count_problems_in_dir()
helper (we don't really need to build a list only to free it at once!),
but it's not that easy because we use get_problem_dirs_for_element_in_time()
which is a public function and I don't want to touch public API (yet?).

Version 2: language fix in the comments

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>
Signed-off-by: Michal Toman <mtoman@redhat.com>